curtin: Test upstream/master with Ubuntu proposed enabled.
Instead of N jobs (one per ubuntu release) running vmtest to test
ubuntu -proposed, just run a single vmtest using master.
The alternative has some value, but I think we get much more bang for
the buck this way.

a.) with the skip_by_date decorator, we'll see if using proposed
actually fixes any issues. If we do SKIP_BY_DATE_BUGS=* we simply skip
those bugs (and that is needed there).

b.) we only have to have one job, not 4.
we get the same coverage with the one job, as it runs curtin installs
of each supported release.  The 4 jobs don't actually change the
coverage at all as the thing intended to be tested here is Ubuntu
-proposed, rather than curtin itself.

c.) we can fix any issue found quickly with fixes to trunk and not have
to wait for SRU.

Also here, enable setting repos in the curtin-vmtest-devel-amd64
test.